deepfake-voice-detector/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
scipy>=1.10.0
webrtcvad-wheels>=2.0.10
noisereduce>=3.0.0
praat-parselmouth>=0.4.3
loguru>=0.7.0
python-multipart>=0.0.6
PyJWT>=2.8.0
websockets>=11.0.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
